Refactor: Split utils module into registry and bin_utils
Split the monolithic `psd_tools.utils` module into two focused modules:

- `psd_tools.registry`: Registry pattern helper (`new_registry` function)
  Located at top-level since it's used across both low-level and high-level APIs

- `psd_tools.psd.bin_utils`: Binary processing utilities (30+ functions)
  Located within psd_tools.psd package as these are low-level binary utilities

The original `psd_tools.utils` module now serves as a deprecation stub that
re-exports from the new modules with a DeprecationWarning to guide users
to migrate their imports.

Changes:
- Created src/psd_tools/registry.py with new_registry function
- Created src/psd_tools/psd/bin_utils.py with all binary utilities
- Updated 19 files to import from new locations
- Moved tests/psd_tools/test_utils.py to tests/psd_tools/psd/test_bin_utils.py
- Added deprecation stub to src/psd_tools/utils.py for backward compatibility

Benefits:
- Clearer architecture and better module organization
- Binary utilities clearly scoped to psd_tools.psd package
- Registry pattern at appropriate abstraction level
- Backward compatibility maintained with deprecation warnings
- All 839 tests passing
