+++ src/psd_tools/psd/adjustments.py
@@ -16,9 +16,8 @@ from psd_tools.psd.base import (
 )
 from psd_tools.psd.descriptor import DescriptorBlock, DescriptorBlock2
 from psd_tools.terminology import Enum, Key
-from psd_tools.utils import (
+from psd_tools.psd.bin_utils import (
     is_readable,
-    new_registry,
     read_fmt,
     read_unicode_string,
     write_bytes,
@@ -26,6 +25,7 @@ from psd_tools.utils import (
     write_padding,
     write_unicode_string,
 )
+from psd_tools.registry import new_registry
 from psd_tools.validators import in_
 
 logger = logging.getLogger(__name__)
