+++ src/psd_tools/psd/descriptor.py
@@ -32,8 +32,7 @@ from psd_tools.psd.base import (
     StringElement,
 )
 from psd_tools.terminology import Enum, Event, Form, Key, Klass, Type, Unit
-from psd_tools.utils import (
-    new_registry,
+from psd_tools.psd.bin_utils import (
     read_fmt,
     read_length_block,
     read_unicode_string,
@@ -44,6 +43,7 @@ from psd_tools.utils import (
     write_padding,
     write_unicode_string,
 )
+from psd_tools.registry import new_registry
 from psd_tools.validators import in_
 
 logger = logging.getLogger(__name__)
