+++ src/psd_tools/psd/image_data.py
@@ -16,7 +16,7 @@ from psd_tools.compression import compress, decompress
 from psd_tools.constants import Compression
 from psd_tools.psd.header import FileHeader
 from psd_tools.psd.base import BaseElement
-from psd_tools.utils import pack, read_fmt, write_bytes, write_fmt
+from psd_tools.psd.bin_utils import pack, read_fmt, write_bytes, write_fmt
 from psd_tools.validators import in_
 
 logger = logging.getLogger(__name__)
