+++ src/psd_tools/psd/image_resources.py
@@ -80,9 +80,8 @@ from psd_tools.psd.base import (
 )
 from psd_tools.psd.color import Color
 from psd_tools.psd.descriptor import DescriptorBlock
-from psd_tools.utils import (
+from psd_tools.psd.bin_utils import (
     is_readable,
-    new_registry,
     read_fmt,
     read_length_block,
     read_pascal_string,
@@ -94,6 +93,7 @@ from psd_tools.utils import (
     write_pascal_string,
     write_unicode_string,
 )
+from psd_tools.registry import new_registry
 from psd_tools.validators import in_
 from psd_tools.version import __version__
