+++ src/psd_tools/psd/linked_layer.py
@@ -11,7 +11,7 @@ from attrs import define, field
 from psd_tools.constants import LinkedLayerType
 from psd_tools.psd.base import BaseElement, ListElement
 from psd_tools.psd.descriptor import DescriptorBlock
-from psd_tools.utils import (
+from psd_tools.psd.bin_utils import (
     is_readable,
     read_fmt,
     read_length_block,
