+++ src/psd_tools/psd/patterns.py
@@ -16,7 +16,7 @@ from attrs import define, field
 from psd_tools.compression import compress, decompress
 from psd_tools.constants import ColorMode, Compression
 from psd_tools.psd.base import BaseElement, ListElement
-from psd_tools.utils import (
+from psd_tools.psd.bin_utils import (
     is_readable,
     read_fmt,
     read_length_block,
